befor trying to move the nest app to the top level

+++ nest-service/src/tasks/tasks.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { BoardsService } from './../boards/boards.service';
+import { Inject, forwardRef, HttpException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -10,30 +11,56 @@ export class TasksService {
   constructor(
     @InjectRepository(Task)
     private tasksRepository: Repository<Task>,
+    @Inject(forwardRef(() => BoardsService))
+    private boardsService: BoardsService
   ) {}
 
-  async create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto, boardId: string) {
+    const board = await this.boardsService.findOne(boardId);
+    if (!board) {
+      throw new HttpException('No board with this ID found', 404);
+    }
     const task = await this.tasksRepository.save(createTaskDto);
     return task;
   }
 
-  async findAll() {
-    const tasks = await this.tasksRepository.find();
+  async findAll(boardId: string) {
+    const board = await this.boardsService.findOne(boardId);
+    if (!board) {
+      throw new HttpException('No board with this ID found', 404);
+    }
+    const tasks = await this.tasksRepository.find({boardId});
     return tasks;
   }
 
-  async findOne(id: string) {
+  async findOne(id: string, boardId: string) {
+    const board = await this.boardsService.findOne(boardId);
+    if (!board) {
+      throw new HttpException('No board with this ID found', 404);
+    }
     const task = await this.tasksRepository.findOne(id);
     return task;
   }
 
-  async update(id: string, updateTaskDto: UpdateTaskDto) {
+  async update(id: string, updateTaskDto: UpdateTaskDto, boardId: string) {
+    const board = await this.boardsService.findOne(boardId);
+    if (!board) {
+      throw new HttpException('No board with this ID found', 404);
+    }
     await this.tasksRepository.update(id, updateTaskDto);
     const task = await this.tasksRepository.findOne(id);
     return task;
   }
 
-  async remove(id: string) {
+  async updateUserId(userId: string, newUserId: string | null) {
+    await this.tasksRepository.update({userId}, {userId: newUserId});
+  };
+
+  async remove(id: string, boardId: string) {
+    const board = await this.boardsService.findOne(boardId);
+    if (!board) {
+      throw new HttpException('No board with this ID found', 404);
+    }
     await this.tasksRepository.delete(id);
     return `Task ${id} has been deleted`;
   }
